# HyperClockCache support for SecondaryCache, with refactoring